+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.9.0] - 2017-12-16
+
+### Changed
+
+- The request handler used to generate the response must implement `Interop\Http\Server\RequestHandlerInterface`. Removed support for callables.
+
+### Removed
+
+- Removed `arguments()` option.
+
 ## [0.8.0] - 2017-11-13
 
 ### Changed
@@ -67,6 +77,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
+[0.9.0]: https://github.com/middlewares/error-handler/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/middlewares/error-handler/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/middlewares/error-handler/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/middlewares/error-handler/compare/v0.5.0...v0.6.0